Fix occasional deadlock when waiting for subdevice command to complete.
Occasionally when sending a command to a subdevice thread, the subdevice thread will complete the command and grab the mutex lock before the calling thread can grab the lock which makes the calling thread wait at the mutex lock.  In this instance the subdevice thread will signal completion and then release the mutex.  At this point the calling thread will then wait indefinitely for a signal that was already sent.  This is fixed here by making the calling thread grab the mutex before sending a command through the pipe.