Requeue buffers even when we are not subscribed to their events.
This fixes an issue where focus would not work after using the camera button to focus.  The issue is somewhat of a race condition where if the camera focus button is depressed, then the auto focus is canceled.  The daemon will tell the isp driver to stop the auto focus stats stream and tell the v4l2 driver to unsubscribe to the events.  The v4l2 driver will drop any events that are not subscribed without informing either the isp driver or userspace.  In this instance the daemon and isp driver will not requeue the buffers.  This patch fixes this by not unsubscribing the events with the v4l2 driver.  The buffer will just be requeued when the event is received.